+++ src/main/fr/fireowls/spaceowls/system/trajectory/SimulateTrajectory.java
@@ -24,7 +24,7 @@ public class SimulateTrajectory extends Trajectory {
 	 * constructeur
 	 * @param corp l'objet qui parcours la trajectoire simulée
 	 */
-	public SimulateTrajectory(SpaceSystem ss, SimuleCorp corp, int vitx,int vity) {
+	public SimulateTrajectory(SpaceSystem ss, SimuleCorp corp, double vitx,double vity) {
 		super(TrajectoryType.SIMULATE);
 		this.ss = ss;
 		this.corp = corp;
